refactor: createRoutes 제거 studyRoutes로 통합
