--- src/routes/createRoutes.js
@@ -1,8 +1,0 @@
-import express from 'express';
-import { createStudy } from '../controllers/createController.js';
-
-const router = express.Router();
-
-router.post('/', createStudy);
-
-export default router;
