+++ src/routes/studyRoutes.js
@@ -1,8 +1,10 @@
 import express from 'express';
-import { getStudies } from '../controllers/studyController.js';
+import { getStudies, createStudy } from '../controllers/studyController.js';
 
 const router = express.Router();
 
 router.get('/', getStudies);
+
+router.post('/', createStudy);
 
 export default router;
